Retry Weaver transactions even when boot not completed
On first boot, Android expects to be able to use Weaver just after it
has started persistent and direct-boot-aware apps such as com.android.se
which provides OMAPI.  There is a race condition where these apps might
not have fully started by the time that Weaver is used, which causes the
Weaver implementation in hardware/nxp/weaver/ to return an error.  There
is actually already a retry loop in this Weaver implementation, but for
some reason it's disabled before boot has completed.  This change makes
the retry loop be run unconditionally, which fixes the problem.

Test: Weaver VTS tescases

Change-Id: Ibe11acd19cf6bdbcfedc29576fe8c963a9dfc04a